pyproject.toml
[project]
name = "di-auto"
version = "0.1.0"
description = "Professional document intelligence automation using GPT Vision for US academic transcript extraction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.91.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.1",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
jupyter = ["jupyter>=1.0.0", "ipykernel>=6.0.0"]
dev = ["pytest>=8.0.0", "black>=25.0.0", "ruff>=0.12.0", "mypy>=1.16.0", "coverage>=7.9.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["controllers"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
